Feat #977: Gracefully handle legacy send for accounts with 0 BTS
